fix NMEA date fiddling.
Three errors were corrected:
1) SECONDS_PER_DAY is assumed to be an integer type.  It was recently
errorously changed to to double.
2) when backfilling dates we need to subtract a day at rolloever,
not addit.
3) If we have a date, then it came with a time.  Don't set the backfill
time to 23:59:59.

Add test coverage for the above.

+++ defs.h
@@ -79,8 +79,8 @@ constexpr double FATHOMS_TO_METERS(double a) { return (a) * 1.8288;};
 constexpr double CELSIUS_TO_FAHRENHEIT(double a) { return (((a) * 1.8) + 32);};
 constexpr double FAHRENHEIT_TO_CELSIUS(double a) { return (((a) - 32) / 1.8);};
 
-constexpr double SECONDS_PER_HOUR = 60L * 60;
-constexpr double SECONDS_PER_DAY = 24L * 60 * 60;
+constexpr long SECONDS_PER_HOUR = 60L * 60;
+constexpr long SECONDS_PER_DAY = 24L * 60 * 60;
 
 /* meters/second to kilometers/hour */
 constexpr double MPS_TO_KPH(double a) { return (a)*SECONDS_PER_HOUR/1000.0;};
